requirements.txt
asgiref==3.5.2
Django==3.2.16
django-cleanup==6.0.0
django-filter==22.1
djangorestframework==3.14.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
